[z2453] Implement new stats system for Health and Mana.
Core now suport health and mana from new stats system.

To make it work you'll need to have DB part also wich is not yet public.

So till that DB will be released health and mana will be prossessed like as usual.

Thank to @X-Savior for making it possible.
Thank to @Schmoozerd for his code improvement.