[z2461] This is used to prevent creatures from pulling too many nearby monsters, for example during certain scripted events.
A good example is Mr. Smite in the Deadmines. There are two guards at the
bottom of his bridge which trigger aggro for him, but if he is missing
this flag he will also pull nearby mobs on the ship, which is not expected
behavior.

(based on commit [12612] - cc137f6)

Signed-off-by: Xfurry <xfurry@scriptdev2.com>

(based on commit [s2101] - 066c3b6)

Signed-off-by: Xfurry <xfurry@scriptdev2.com>